+++ src/cljs/rems/catalogue.cljs
@@ -100,13 +100,12 @@
       [:div
        [:h2 (text :t.catalogue/continue-existing-application)]
        (text :t.catalogue/continue-existing-application-intro)
-       [application-list/component
+       [application-list/list
         {:id applications
          :applications applications
          :visible-columns #{:resource :last-activity :view}
          :default-sort-column :last-activity
-         :default-sort-order :desc
-         :filterable? false}]])))
+         :default-sort-order :desc}]])))
 
 (defn- catalogue-table []
   (let [catalogue {:id ::catalogue