+++ test/clj/rems/application/test_search.clj
@@ -54,9 +54,6 @@
                            :actor "alice"
                            :field-values [{:field 1
                                            :value "Supercalifragilisticexpialidocious"}]})
-      (test-data/command! {:type :application.command/submit ; make sure that the required field was filled in
-                           :application-id app-id
-                           :actor "alice"})
       (is (= #{app-id} (search/find-applications "Supercalifragilisticexpialidocious")) "any field")
       (is (= #{app-id} (search/find-applications "title:Supercalifragilisticexpialidocious")) "title field")))
 
@@ -94,11 +91,36 @@
                            :actor "alice"
                            :field-values [{:field 1
                                            :value "Tis but a scratch."}]})
-      (test-data/command! {:type :application.command/submit ; make sure that the required field was filled in
-                           :application-id app-id
-                           :actor "alice"})
       (is (= #{app-id} (search/find-applications "scratch")) "any field")
       (is (= #{app-id} (search/find-applications "form:scratch")) "form field")))
+
+  (testing "updating applications"
+    (let [form-id (test-data/create-form! {:form/fields [{:field/type :text
+                                                          :field/title {:en "Text field"}
+                                                          :field/optional false}]})
+          cat-id (test-data/create-catalogue-item! {:form-id form-id})
+          app-id (test-data/create-application! {:catalogue-item-ids [cat-id]
+                                                 :actor "alice"})]
+      (test-data/command! {:type :application.command/save-draft
+                           :application-id app-id
+                           :actor "alice"
+                           :field-values [{:field 1
+                                           :value "version1"}]})
+      (is (= #{app-id} (search/find-applications "version1"))
+          "original version is indexed")
+
+      (test-data/command! {:type :application.command/save-draft
+                           :application-id app-id
+                           :actor "alice"
+                           :field-values [{:field 1
+                                           :value "version2"}]})
+      (is (= #{} (search/find-applications "version1"))
+          "should not find old versions")
+      (is (= #{app-id} (search/find-applications "version2"))
+          "should find the new version")))
+
+  (testing "multiple results"
+    (is (< 1 (count (search/find-applications "alice")))))
 
   (testing "invalid query"
     (is (= nil (search/find-applications "+")))))